Move heavy single-purpose deps into optional extras
ord-schema pulled several large dependencies into every install for
features most consumers never touch. Split them into extras so the
default footprint is just the schema + core helpers:

  - huggingface (huggingface-hub): extract fetch_dataset out of
    message_helpers into a new ord_schema/huggingface.py that imports
    huggingface_hub at module top. message_helpers (imported by ~12
    modules) is now HF-free; importing ord_schema.huggingface without the
    extra fails naturally. Drops typer/fsspec/hf-xet from default installs.
  - orm (sqlalchemy, psycopg, inflection): the ord_schema.orm subpackage
    is self-contained and nothing in core imports it, so moving these
    needs no code change -- importing ord_schema.orm requires [orm].
  - github (pygithub): only process_dataset.py uses it (to post a
    submission summary to a GitHub issue); requires [github] to run.

CI installs all extras; uv.lock keeps the packages (now grouped under the
extras). Full suite (544 tests) green with ruff + ty clean.
