+++ .github/workflows/run_tests.yml
@@ -48,7 +48,7 @@ jobs:
           enable-cache: true
           python-version: ${{ matrix.python-version }}
       - name: Install ord_schema
-        run: uv sync --extra tests --locked
+        run: uv sync --extra github --extra huggingface --extra orm --extra tests --locked
       - name: Run tests
         shell: bash -l {0}
         # Exercise the live PubChem/OPSIN resolver smoke tests on a single matrix
