+++ ord_schema/message_helpers.py
@@ -30,8 +30,6 @@ from google.protobuf import (
     text_format,
 )
 from google.protobuf.message import DecodeError
-from huggingface_hub import hf_hub_download
-from huggingface_hub.errors import EntryNotFoundError
 from rdkit import Chem
 from rdkit.Chem import rdChemReactions
 
@@ -45,9 +43,6 @@ _COMPOUND_IDENTIFIER_LOADERS = {
     reaction_pb2.CompoundIdentifier.MOLBLOCK: Chem.MolFromMolBlock,
 }
 MessageType = TypeVar("MessageType")  # Generic for setting return types
-# Hugging Face dataset that mirrors the ord-data repository; see
-# https://huggingface.co/datasets/open-reaction-database/ord-data.
-ORD_DATA_HF_REPO = "open-reaction-database/ord-data"
 
 
 def build_compound(
@@ -807,66 +802,6 @@ class MessageFormat(enum.Enum):
 
 _BINARY_FORMATS = {MessageFormat.BINARY, MessageFormat.BINPB}
 _TEXT_FORMATS = {MessageFormat.PBTXT, MessageFormat.TXTPB}
-
-
-def fetch_dataset(
-    dataset_id: str,
-    *,
-    revision: str = "main",
-    cache_dir: str | None = None,
-    local_dir: str | None = None,
-) -> str:
-    """Downloads a dataset file from the ord-data Hugging Face mirror.
-
-    Prefers the Parquet serialization and falls back to the legacy ``.pb.gz``
-    binary format for datasets that have not been converted yet. The file is
-    not parsed; dispatch on the returned suffix to choose a reader, e.g.
-    ``parquet_dataset.read_dataset``/``parquet_dataset.DatasetView`` for
-    ``.parquet`` and ``load_message`` for ``.pb.gz``.
-
-    Downloads are cached and content-verified by ``huggingface_hub``: repeat
-    calls reuse the cached file (re-validating against the remote for moving
-    revisions like ``"main"``) instead of re-downloading. By default the cache
-    lives under ``~/.cache/huggingface/hub``; set ``$HF_HOME`` (or pass
-    ``cache_dir``) to relocate it, or set ``$HF_HUB_OFFLINE=1`` to reuse an
-    already-warm cache without network access.
-
-    Args:
-        dataset_id: Dataset ID.
-        revision: Branch, tag, or commit SHA to download (default: ``"main"``).
-        cache_dir: Override the Hugging Face cache directory. Files are stored
-            content-addressed and returned via a symlink into the cache.
-        local_dir: If set, materialize the file as a real (non-symlink) copy
-            under this directory instead of returning a cache symlink; useful
-            for staging files into a project tree.
-
-    Returns:
-        Local filesystem path to the downloaded dataset file.
-
-    Raises:
-        RuntimeError: If no file exists for the dataset.
-        ValueError: If the dataset ID is invalid.
-    """
-    # Lazy import breaks a real circular dependency.
-    from ord_schema import validations  # noqa: PLC0415
-
-    if not validations.is_valid_dataset_id(dataset_id):
-        raise ValueError(f"Invalid dataset ID: {dataset_id}")
-    for suffix in (".parquet", ".pb.gz"):
-        try:
-            return hf_hub_download(
-                repo_id=ORD_DATA_HF_REPO,
-                repo_type="dataset",
-                revision=revision,
-                filename=id_filename(f"{dataset_id}{suffix}"),
-                cache_dir=cache_dir,
-                local_dir=local_dir,
-            )
-        except EntryNotFoundError:
-            continue
-    raise RuntimeError(
-        f"Dataset {dataset_id} not found in {ORD_DATA_HF_REPO}@{revision}"
-    )
 
 
 def _message_format(path: pathlib.Path) -> MessageFormat:
